Core/Datastores: corrected a member type in AchievementCriteriaEntry

+++ src/server/game/DataStores/DBCStructure.h
@@ -166,7 +166,7 @@ struct AchievementCriteriaEntry
         uint32 GuildChallengeType;
     } Asset;
 
-    uint32 Amount;
+    uint64 Amount;
 
     struct
     {
